fix modal and change date picker to textfield component
