+++ src/components/Teams/NewTeam.jsx
@@ -2,44 +2,30 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useNavigate } from 'react-router-dom';
+import { useForm } from 'react-hook-form';
 import {
-  Box, Button, InputAdornment, MenuItem, OutlinedInput, TextField, useTheme,
+  Box, Button, MenuItem, OutlinedInput, TextField, useTheme,
 } from '@mui/material';
-import { Controller, useForm } from 'react-hook-form';
-// import { DatePicker, LocalizationProvider } from '@mui/x-date-pickers';
-import { PhotoCamera } from '@mui/icons-material';
-// import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
-import { createMember, getChurches } from '../../redux/thunk';
-import styles from '../../styles/Members.module.css';
-import TeamsDropdown from '../Teams/TeamsDropdown';
+import { createTeam, getChurches } from '../../redux/thunk';
 import { tokens } from '../../theme';
 
-const NewMember = () => {
+const NewTeam = () => {
   const [msg, setMsg] = useState('');
-  const [selectedTeams, setSelectedTeams] = useState([]);
   const [selectedChurch, setSelectedChurch] = useState([]);
-  const [fileImg, setFileImg] = useState('');
 
-  const {
-    churches, isLoading,
-  } = useSelector((store) => store.churches);
-
-  const navigate = useNavigate();
   const dispatch = useDispatch();
+  const form = useForm();
+  const {
+    register, handleSubmit, reset, formState,
+  } = form;
+  const { errors } = formState;
 
   const theme = useTheme();
   const colors = tokens(theme.palette.mode);
 
-  const form = useForm();
   const {
-    register, handleSubmit, reset, formState, control,
-  } = form;
-  const { errors } = formState;
-
-  const handleImgUpload = (e) => {
-    setFileImg(URL.createObjectURL(e.target.files[0]));
-  };
+    churches, isLoading,
+  } = useSelector((store) => store.churches);
 
   const handleChurchSelect = (event) => {
     const {
@@ -52,30 +38,20 @@ const NewMember = () => {
   };
 
   const onSubmit = (data) => {
-    const member = new FormData();
-    member.append('member[name]', data.name);
-    member.append('member[photo]', data.photo[0]);
-    member.append('member[address]', data.address);
-    member.append('member[phone_number]', data.phone_number);
-    member.append('member[joined_at]', data.joined_at);
-    member.append('member[church_id]', data.church_id);
-    if (selectedTeams.length > 0) {
-      selectedTeams.forEach((team) => {
-        member.append('member[teams][]', team.id);
-      });
-    } else {
-      member.append('member[teams]', '');
-    }
-    dispatch(createMember(member)).then(setMsg('Member Added Successfully!'));
+    const team = new FormData();
+    team.append('team[name]', data.name);
+    team.append('team[main_leader_name]', data.main_leader_name);
+    team.append('team[sub_leader_name]', data.sub_leader_name);
+    team.append('team[established_at]', data.established_at);
+    team.append('team[church_id]', data.church_id);
+
+    dispatch(createTeam(team)).then(setMsg('Team Added Successfully'));
     setTimeout(() => {
       setMsg('');
     }, 3000);
 
     reset();
-    setFileImg('');
-    navigate('/members');
   };
-
   useEffect(() => {
     dispatch(getChurches());
   }, [dispatch]);
@@ -92,6 +68,7 @@ const NewMember = () => {
           gap="4%"
           overflow="auto"
         >
+
           <TextField
             label="name"
             type="text"
@@ -118,48 +95,19 @@ const NewMember = () => {
               width: '48%',
             }}
           />
-          <Box
-            display="flex"
-            sx={{
-              width: '48%',
-            }}
-          >
-            <TextField
-            // label="photo"
-              type="file"
-              variant="outlined"
-              InputProps={{
-                startAdornment: (
-                  <InputAdornment position="start">
-                    <PhotoCamera />
-                  </InputAdornment>
-                ),
-              }}
-              {...register('photo', { required: 'Photo is required' })}
-              onChange={handleImgUpload}
-              error={Boolean(errors.photo)}
-              helperText={errors.photo?.message}
-              margin="normal"
-              sx={{
-                '& label.Mui-focused': {
-                  color: colors.orangeAccent[500],
-                },
-                '& .MuiOutlinedInput-root': {
-                  '&.Mui-focused fieldset': {
-                    borderColor: colors.orangeAccent[500],
-                  },
-                },
-                mr: fileImg === '' ? '' : 1,
-              }}
-            />
-            {fileImg === '' ? '' : <img src={fileImg} alt="member" className={styles.memberDetailImg} />}
-          </Box>
           <TextField
-            label="address"
+            label="Main Leader's Name"
             type="text"
-            {...register('address', { required: 'Address is required' })}
-            error={Boolean(errors.address)}
-            helperText={errors.address?.message}
+            {...register('main_leader_name', {
+              required:
+                  {
+                    value: true,
+                    message: 'Main leader&apos;s name is required',
+                  },
+            })
+              }
+            error={Boolean(errors.main_leader_name)}
+            helperText={errors.main_leader_name?.message}
             margin="normal"
             sx={{
               '& label.Mui-focused': {
@@ -171,19 +119,21 @@ const NewMember = () => {
                 },
               },
               width: '48%',
-              mt: 1,
             }}
           />
           <TextField
-            label="phone number"
+            label="Sub Leader's Name"
             type="text"
-            {...register('phone_number', {
-              required: 'Phone number is required',
-              pattern:
-                { value: /^[0-9]*$/, message: 'Add correct phone number' },
-            })}
-            error={Boolean(errors.phone_number)}
-            helperText={errors.phone_number?.message}
+            {...register('sub_leader_name', {
+              required:
+                  {
+                    value: true,
+                    message: 'Main leader&apos;s name is required',
+                  },
+            })
+              }
+            error={Boolean(errors.sub_leader_name)}
+            helperText={errors.sub_leader_name?.message}
             margin="normal"
             sx={{
               '& label.Mui-focused': {
@@ -195,17 +145,14 @@ const NewMember = () => {
                 },
               },
               width: '48%',
-              mt: 1,
             }}
           />
           <TextField
             type="date"
-            {...register('joined_at', {
-              required: 'Date is required',
-            })}
+            {...register('established_at', { required: 'Established date is required' })}
             defaultValue=""
-            error={Boolean(errors.joined_at)}
-            helperText={errors.joined_at?.message}
+            error={Boolean(errors.established_at)}
+            helperText={errors.established_at?.message}
             margin="normal"
             sx={{
               '& label.Mui-focused': {
@@ -217,7 +164,6 @@ const NewMember = () => {
                 },
               },
               width: '48%',
-              mt: 1,
               '& ::-webkit-calendar-picker-indicator': {
                 filter: 'invert(1)',
               },
@@ -267,27 +213,7 @@ const NewMember = () => {
                 </MenuItem>
               ))}
           </TextField>
-          <Box
-            width="100%"
-            sx={{ mt: 2 }}
-          >
-            <Controller
-              control={control}
-              name="teams"
-              render={({ field }) => (
 
-                <TeamsDropdown
-                  field={field}
-                  defaultValue={[]}
-                  selectedTeams={selectedTeams}
-                  setSelectedTeams={setSelectedTeams}
-                />
-
-              )}
-            />
-            <span className="errorMsg">{errors.teams?.message}</span>
-
-          </Box>
           <Button
             type="submit"
             variant="contained"
@@ -295,7 +221,7 @@ const NewMember = () => {
             fullWidth
             sx={{ background: colors.greenAccent[700], ':hover': { background: colors.greenAccent[600] }, mt: 2 }}
           >
-            Add Member
+            Add Team
           </Button>
           <span>{msg}</span>
         </Box>
@@ -304,4 +230,4 @@ const NewMember = () => {
   );
 };
 
-export default NewMember;
+export default NewTeam;
